Split DeviceSelector/index.tsx into component files
