+++ streamlit_webrtc/frontend/src/DeviceSelector/types.ts
@@ -1,0 +1,4 @@
+export interface DevicesMap {
+  audio: MediaDeviceInfo[];
+  video: MediaDeviceInfo[];
+}
